requirements.txt
pytz
requests
gunicorn
python-dotenv
python-decouple
django
djangorestframework
django-timezone-field 
django-cors-headers
uvicorn
httptools
psycopg2-binary
whitenoise
celery
redis
django-celery-beat
aiohttp
pyrofork
TgCrypto
python-telegram-bot
django-redis
